Move everything up to the root directory for this branch
