--- example/.gitignore
@@ -1,2 +1,0 @@
-/node_modules
-/build
