--- example/bad-example.ts
@@ -1,9 +1,0 @@
-export interface NotExternalInterface {
-    firstName: string;
-    lastName: string;
-}
-
-// // Uncommenting this makes it's conflict
-// export interface ExternalInterface {
-//     phone: string[];
-// }
