+++ example/interfaces.ts
@@ -1,37 +1,26 @@
-export enum Permission {
-    Admin = 'admin',
-    User = 'user',
-    Anonymous = 'anonymous'
+export enum CertificateType {
+    Application = 'application',
+    User = 'user'
 }
 
-// Other comment
-// @runtime
-export interface ExternalInterface {
-    user: {
-        firstName: string;
-        lastName: string;
-        phoneNumber?: string;
-        id: number;
-    };
-    permission: Permission;
+export enum NonInitializedEnum {
+    TWELVE,
+    A = '5',
+    B = 4,
+    A1,
+    C
 }
 
-export interface AlsoExternalInterface {
-    firstName: string;
-    lastName: string;
-    phoneNumber?: string;
-}
+type TypeAlias = CertificateType[];
 
-// @runtime
-export type CustomName = AlsoExternalInterface;
-
-// @runtime
-export interface References extends AlsoExternalInterface {
-    tacos: boolean;
-    ref: AlsoExternalInterface;
-}
-
-export interface YetAnotherExternalInterface {
-    firstName: string;
-    phoneNumber: string;
+export interface RecordKey {
+//     key: string;
+    num: number[];
+    someKeyName: TypeAlias;
+    // nestedType: { nestedNumber?: number; nestedCert: CertificateType };
+    // key2: string | number | CertificateType;
+    // certType: CertificateType;
+    // wat: NonInitializedEnum;
+    // extras?: string[];
+    // certs: CertificateType[];
 }
