--- example/output.d.ts
@@ -1,24 +1,0 @@
-// prettier-ignore-start
-// tslint:disable
-/*
- * This file was automatically generated by something else. Probably shouldn't touch it.
- */
-import * as _unused from "ts-runtime-interface";
-import { ExternalInterface } from "./interfaces";
-import { CustomName } from "./interfaces";
-import { References } from "./interfaces";
-
-declare namespace Validate {
-  function validate(
-    interfaceName: "ExternalInterface",
-    obj: any
-  ): ExternalInterface;
-  function validate(interfaceName: "CustomName", obj: any): CustomName;
-  function validate(interfaceName: "References", obj: any): References;
-  function validate(interfaceName: string, obj: any): never;
-}
-
-declare module "ts-runtime-interface" {
-  // @ts-ignore
-  export function validator(schemaPath: string): typeof Validate.validate;
-}
