+++ example/usage.ts
@@ -1,44 +1,16 @@
-import { validator } from 'ts-runtime-interface';
-import { assertError, assertEqual } from './assert';
-import { Permission } from './interfaces';
+import { validateInterface as renamed } from 'runtime-check';
 
-const validate = validator('./runtimeSchema.json');
+import { RecordKey } from './interfaces';
 
-const parsedExternalInterface = JSON.parse(`{
-    "user": {
-        "firstName": "Jacob",
-        "lastName": "Gardner",
-        "id": 12
-    },
-    "permission": "anonymous"
-}`);
-const parsedCustomInterface = JSON.parse(`{
-    "firstName": "Jacob",
-    "lastName": "Gardner",
-    "extraField": 182301,
-    "phoneNumber": "1238"
-}`);
+// type Dup = RecordKey;
 
-parsedExternalInterface; // $ExpectType any
+const key: RecordKey = renamed({ key: 'thing', certType: 'application' }); // $ExpectType RecordKey
+const key2 = renamed({ key: 'thing', certType: 'application' }) as RecordKey; // $ExpectType RecordKey
 
-const validated = validate('ExternalInterface', parsedExternalInterface); // $ExpectType ExternalInterface
-assertEqual(validated.permission, Permission.Anonymous);
-assertEqual(validated.user.firstName, 'Jacob');
-assertEqual(validated.user.lastName, 'Gardner');
-assertEqual(validated.user.id, 12);
-assertEqual(validated.user.phoneNumber, undefined);
+// let delayed: RecordKey;
 
-const custom = validate('CustomName', parsedCustomInterface); // $ExpectType AlsoExternalInterface
-assertEqual(custom.firstName, 'Jacob');
-assertEqual(custom.lastName, 'Gardner');
-assertEqual(custom.phoneNumber, '1238');
+// delayed = renamed({ key: 'thing', certType: 'application' }); // $ExpectType RecordKey
 
-assertError(() => {
-    validate('UnknownInterface', parsedExternalInterface); // $ExpectType never
-});
+// let delayedType;
+// delayedType = renamed({ key: 'thing', certType: 'application' }) as Dup; // $ExpectType RecordKey
 
-assertError(() => {
-    const { firstName, ...obj } = parsedCustomInterface;
-
-    validate('CustomName', obj); // $ExpectType AlsoExternalInterface
-});
