--- lib/tsconfig.json
@@ -1,8 +1,0 @@
-{
-    "extends": "../tsconfig.json",
-    "compilerOptions": {
-        "outDir": "./build",
-        "declaration": true
-    },
-    "exclude": ["build", "type-example-overrides.d.ts"]
-}
