+++ package.json
@@ -1,52 +1,25 @@
 {
-    "name": "ts-runtime-interface",
-    "version": "0.1.3",
-    "description": "",
-    "main": "./lib/build/lib.js",
-    "scripts": {
-        "build": "tsc -p ./lib",
-        "prepare": "npm run build",
-        "process": "node ./lib/build/main.js -p tsconfig.json ./output.d.ts ./runtimeSchema.json"
-    },
-    "bin": {
-        "ts-runtime-interface": "./lib/build/main.js"
-    },
-    "repository": {
-        "type": "git",
-        "url": "https://github.com/jacobgardner/ts-runtime-interface.git"
-    },
-    "homepage": "https://github.com/jacobgardner/ts-runtime-interface",
-    "files": [
-        "lib/build/"
-    ],
-    "author": "Jacob Gardner <jacob.v.gardner@gmail.com>",
-    "license": "SEE LICENSE IN LICENSE",
-    "keywords": [
-        "typescript",
-        "wins",
-        "innovate",
-        "redefine",
-        "jabberwocky"
-    ],
-    "dependencies": {
-        "commander": "^2.19.0",
-        "js-yaml": "^3.13.1",
-        "jsonschema": "^1.2.4",
-        "mkdirp": "^0.5.1",
-        "prettier": "^1.16.4",
-        "true-case-path": "^1.0.3",
-        "typescript-json-schema": "^0.38.3"
-    },
-    "devDependencies": {
-        "@types/jsonschema": "^1.1.1",
-        "@types/mkdirp": "^0.5.2",
-        "@types/node": "^11.13.14",
-        "@types/prettier": "^1.16.4",
-        "dtslint": "^0.5.1",
-        "tslint-config-prettier": "^1.18.0",
-        "typescript": "^3.5.1"
-    },
-    "peerDependencies": {
-        "typescript": "^3.5.1"
-    }
+  "name": "runtime-check",
+  "version": "0.1.0",
+  "description": "",
+  "main": "build/main.js",
+  "types": "./index.d.ts",
+  "scripts": {
+    "check": "npm run build -- --noEmit",
+    "build": "tsc",
+    "watch": "npm run build -- --watch"
+  },
+  "author": "",
+  "license": "ISC",
+  "devDependencies": {
+    "@types/node": "^12.0.8",
+    "typescript": "^3.5.1"
+  },
+  "dependencies": {
+    "@types/colors": "^1.2.1",
+    "@types/json-schema": "^7.0.3",
+    "colors": "^1.3.3",
+    "json-schema": "^0.2.3",
+    "ts-json-schema-generator": "^0.46.0"
+  }
 }
