--- schema.json
@@ -1,1 +1,0 @@
-{"$schema":"http://json-schema.org/draft-07/schema#","definitions":{"ExternalInterface":{"type":"object","properties":{"user":{"type":"object","properties":{"firstName":{"type":"string"},"lastName":{"type":"string"},"phoneNumber":{"type":"string"},"id":{"type":"number"}},"required":["firstName","id","lastName"]},"permission":{"$ref":"#/definitions/Permission"}},"required":["permission","user"]},"CustomName":{"type":"object","properties":{"firstName":{"type":"string"},"lastName":{"type":"string"},"phoneNumber":{"type":"string"}},"required":["firstName","lastName"]},"References":{"type":"object","properties":{"tacos":{"type":"boolean"},"ref":{"$ref":"#/definitions/AlsoExternalInterface"},"firstName":{"type":"string"},"lastName":{"type":"string"},"phoneNumber":{"type":"string"}},"required":["firstName","lastName","ref","tacos"]},"Permission":{"enum":["admin","anonymous","user"],"type":"string"},"AlsoExternalInterface":{"type":"object","properties":{"firstName":{"type":"string"},"lastName":{"type":"string"},"phoneNumber":{"type":"string"}},"required":["firstName","lastName"]}}}
