--- v2/example/package.json
@@ -1,20 +1,0 @@
-{
-  "name": "example",
-  "version": "1.0.0",
-  "description": "",
-  "main": "index.js",
-  "scripts": {
-    "check": "npm run build -- --noEmit",
-    "build": "ttsc usage.ts",
-    "build:errors": "ttsc errors.ts",
-    "watch": "npm run build -- --watch"
-  },
-  "author": "",
-  "private": true,
-  "license": "ISC",
-  "devDependencies": {
-    "@types/node": "^12.0.8",
-    "ttypescript": "^1.5.7",
-    "typescript": "^3.5.1"
-  }
-}
