--- v2/package.json
@@ -1,25 +1,0 @@
-{
-  "name": "runtime-check",
-  "version": "0.1.0",
-  "description": "",
-  "main": "build/main.js",
-  "types": "./index.d.ts",
-  "scripts": {
-    "check": "npm run build -- --noEmit",
-    "build": "tsc",
-    "watch": "npm run build -- --watch"
-  },
-  "author": "",
-  "license": "ISC",
-  "devDependencies": {
-    "@types/node": "^12.0.8",
-    "typescript": "^3.5.1"
-  },
-  "dependencies": {
-    "@types/colors": "^1.2.1",
-    "@types/json-schema": "^7.0.3",
-    "colors": "^1.3.3",
-    "json-schema": "^0.2.3",
-    "ts-json-schema-generator": "^0.46.0"
-  }
-}
